+++ gitman/commands.py
@@ -58,7 +58,8 @@ def install(*names, root=None, depth=None,
     - `*names`: optional list of dependency directory names to filter on
     - `root`: specifies the path to the root working tree
     - `depth`: number of levels of dependencies to traverse
-    - `force`: indicates uncommitted changes can be overwritten
+    - `force`: indicates uncommitted changes can be overwritten and
+               script errors can be ignored
     - `fetch`: indicates the latest branches should always be fetched
     - `clean`: indicates untracked files should be deleted from dependencies
 
@@ -72,13 +73,16 @@ def install(*names, root=None, depth=None,
     config = load_config(root)
 
     if config:
-        common.show()
+        common.newline()
         common.show("Installing dependencies...", color='message', log=False)
-        common.show()
+        common.newline()
         count = config.install_dependencies(
             *names, update=False, depth=depth,
             force=force, fetch=fetch, clean=clean,
         )
+
+        if count:
+            _run_scripts(*names, depth=depth, force=force, _config=config)
 
     return _display_result("install", "Installed", count)
 
@@ -94,7 +98,8 @@ def update(*names, root=None, depth=None,
     - `root`: specifies the path to the root working tree
     - `depth`: number of levels of dependencies to traverse
     - `recurse`: indicates nested dependencies should also be updated
-    - `force`: indicates uncommitted changes can be overwritten
+    - `force`: indicates uncommitted changes can be overwritten and
+               script errors can be ignored
     - `clean`: indicates untracked files should be deleted from dependencies
     - `lock`: indicates actual dependency versions should be recorded
 
@@ -109,21 +114,41 @@ def update(*names, root=None, depth=None,
     config = load_config(root)
 
     if config:
-        common.show()
+        common.newline()
         common.show("Updating dependencies...", color='message', log=False)
-        common.show()
+        common.newline()
         count = config.install_dependencies(
             *names, update=True, depth=depth,
             recurse=recurse, force=force, fetch=True, clean=clean,
         )
-        common.dedent(level=0)
+
         if count and lock is not False:
             common.show("Recording installed versions...",
                         color='message', log=False)
-            common.show()
+            common.newline()
             config.lock_dependencies(*names, obey_existing=lock is None)
 
+        if count:
+            _run_scripts(*names, depth=depth, force=force, _config=config)
+
     return _display_result("update", "Updated", count)
+
+
+def _run_scripts(*names, depth=None, force=False, _config=None):
+    """Run post-install scripts.
+
+    Optional arguments:
+
+    - `*names`: optional list of dependency directory names filter on
+    - `depth`: number of levels of dependencies to traverse
+    - `force`: indicates script errors can be ignored
+
+    """
+    assert _config, "'_config' is required"
+
+    common.show("Running scripts...", color='message', log=False)
+    common.newline()
+    _config.run_scripts(*names, depth=depth, force=force)
 
 
 @restore_cwd
@@ -144,10 +169,10 @@ def display(*, root=None, depth=None, allow_dirty=True):
     config = load_config(root)
 
     if config:
-        common.show()
+        common.newline()
         common.show("Displaying current dependency versions...",
                     color='message', log=False)
-        common.show()
+        common.newline()
         config.log(datetime.datetime.now().strftime("%F %T"))
         count = 0
         for identity in config.get_dependencies(depth=depth,
@@ -176,9 +201,9 @@ def lock(*names, root=None):
     config = load_config(root)
 
     if config:
-        common.show()
+        common.newline()
         common.show("Locking dependencies...", color='message', log=False)
-        common.show()
+        common.newline()
         count = config.lock_dependencies(*names, obey_existing=False)
         common.dedent(level=0)
 
@@ -202,14 +227,14 @@ def delete(*, root=None, force=False):
     config = load_config(root)
 
     if config:
-        common.show()
+        common.newline()
         common.show("Checking for uncommitted changes...",
                     color='message', log=False)
-        common.show()
+        common.newline()
         count = len(list(config.get_dependencies(allow_dirty=force)))
         common.dedent(level=0)
         common.show("Deleting all dependencies...", color='message', log=False)
-        common.show()
+        common.newline()
         config.uninstall_dependencies()
 
     return _display_result("delete", "Deleted", count, allow_zero=True)